程序查询页面通过

+++ src/main/java/com/infoa/educationms/DTO/GradeDTO.java
@@ -6,7 +6,7 @@ public class GradeDTO {
     private Integer gradeId;
     private Integer takesId;
     private Integer grade;
-    private Double proportion;
+    private Float proportion;
     private String type; // 'attending', 'homework', 'test'
     private Integer studentId;
     private String studentName;
